+++ test/test1.sh
@@ -87,7 +87,7 @@ expectfn "$clixon_cli -1f $clixon_cf -l o debug level 1" ""
 expectfn "$clixon_cli -1f $clixon_cf -l o debug level 0" ""
 
 new "cli downcall"
-expectfn "$clixon_cli -1f $clixon_cf -l o downcall \"This is a test =====\"" "^\"This is a test =====\"$"
+expectfn "$clixon_cli -1f $clixon_cf -l o rpc ipv4" "^<rpc-reply>"
 
 new "Kill backend"
 # Check if still alive